deduplicate code in unittests and fix doxygen
