+++ doc/doxygen/doxygen.conf
@@ -791,7 +791,8 @@ WARN_LOGFILE           =
 # spaces. See also FILE_PATTERNS and EXTENSION_MAPPING
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = "../../src" \
+INPUT                  = "../../include/nfm" \
+                         "../../src" \
                          "doxygen_stl.cpp"
 
 # This tag can be used to specify the character encoding of the source files
